fix(replayer): Fix null objects
These objects can all be null (e.g. if player is playing and we navigate to a page without player, due to async, the player can be destroyed by the time we try to access these objects). Remove the non-null assertion operator as typescript is right that they can be null.