fix(replayer): Fix null objects on playback (#38)
These objects can all be null (e.g. if player is playing and we navigate
to a page without player, due to async, the player can be destroyed by
the time we try to access these objects). Remove the non-null assertion
operator as typescript is right that they can be null.

Fixes https://sentry.sentry.io/issues/2628382889/,
https://sentry.sentry.io/issues/3616367681/,
https://sentry.sentry.io/issues/2669497075/,
https://sentry.sentry.io/issues/3549471392/