fix: Ensure we do not check `window` in module scope (#47)
It appears the new check fails on node, even when not importing rrweb. I
guess it may be because in development it is not treeshaken, and module
scope is still evaluated.

Now, we check this every time. This may be slightly more expensive, but
also safer as we'd catch it if this is removed/overwritten later on.