feat: Ensure password inputs are always masked (#78)
This PR does two things:

1. Ensure you cannot opt-out of masking `type="password"` inputs
2. Ensure inputs that used to have `type=password` and have this type
changed (so e.g. through a "show password" toggle) always keep the
password masked, even afterwards.

We should never record passwords, and e.g. `toggle password` type
buttons are quite common and would easily leak the password into the
replay as of now.

We do this by adding a `rr_is_password` attribute to the HTML input when
the type is changed from `password` to something else. This is IMHO the
easiest solution for this, and since we only add this when the type is
really changed (so not eagerly for all inputs) IMHO it's an acceptable
tradeoff to have this in the DOM.

Closes https://github.com/getsentry/rrweb/issues/34