feat: Use JS to disable clicks in Replayer (#54)
We turned `pointer-events` on in our client so that our users can use
browser tools to "inspect" the dom. Without `pointer-events`, users
hovering over elements in the replay would not reflect in the dev tools.

A side-effect of enabling `pointer-events` was that users would be able
to click links in the Replay and end up breaking everything. e.g. "Looks
like your Replayer was destroyed"