chore: Include original rrweb and Sentry licenses (#92)
While setting up the publishing process for our `rrweb` fork, we changed
the `LICENSE` text (#18) to our standard MIT license. This was recently
pointed out in https://github.com/rrweb-io/rrweb/issues/1248. While It's
worth noting that our fork was also always MIT-licensed, we did changed
the copyright, which incorrectly suggested that the repo's code was
ours. While I'm not entirely sure what the reason for the change was
back then, in hindsight this shouldn't have happened and I'm sorry for
causing any confusion associated with this change.

This patch adjusts the `LICENSE` file to include both, rrweb's original
license as well as our license text, which is what we settled on after
internal conversations (see
https://github.com/getsentry/rrweb/pull/91#issuecomment-1623839222).

---------

Co-authored-by: Chad Whitacre <chad@zetaweb.com>