feat: Allow for masking of attributes [WIP]
Note: this is more of an RFC

This currently adds a new API, `maskAttributesFn: (key: string, value: string) => string`, that is used as a callback in `transformAttribute`. I prefer this API as it gives more flexibility for users (though it may need to pass the el node for most flexibility), but it is inconsistent with `maskTextFn` and `maskInputFn`.

other options:

* Rename this to something else (open to ideas)
* Change this to pass value, and dom element (similar to MaskInputFn) to customize masking instead of decision maker of when to mask and introduce a simpler declarative API for what attributes to mask
* ???