build: Build for ES2020 (#142)
Avoid already polyfilling stuff in here - we rather do this in the
Sentry SDK directly.

Currently, we are polyfilling things separately here (e.g. rest
parameters, async/await), but we would polyfill stuff we need more
efficiently in the Sentry SDK anyhow. So this should reduce bundle size
a bit.